[#38] refactor: Change text location to atoms
