+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as AppBootstrap } from "./app-bootstrap/app-bootstrap";
-export { default as Text } from "./text/text";
+export { default as Text } from "./atoms/text/text";
 export { default as ButtonNormal } from "./atoms/button/buttonNormal";
 export { default as ButtonGradient } from "./atoms/button/buttonGradient";
 export { default as Modal } from "./modal/modal";
