+++ src/components/organisms/modal/modal.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import { View } from "react-native";
-import Text from "../../text/text";
+import Text from "../../atoms/text/text";
 import RNModal from "react-native-modal";
 import styles from "./modal.style";
