+++ src/components/record-list/record-list.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import { View, FlatList, Image, TouchableOpacity } from "react-native";
-import Text from "../text/text";
+import Text from "../atoms/text/text";
 import styles from "./record-list.styles";
 
 // calender 양 옆에 마진 값을 줘야 함... 헤더 스타일 고치고, 어떻게 하면 마킹할 수 있는지 알아보기
